// clients/js/src/generated/pdas/receipt.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  fixEncoderSize,
  getAddressEncoder,
  getBytesEncoder,
  getProgramDerivedAddress,
  type Address,
  type ProgramDerivedAddress,
  type ReadonlyUint8Array,
} from '@solana/kit';

export type ReceiptSeeds = {
  receiver: Address;
  mint: Address;
  sourceOwner: Address;
  uniqueNonce: ReadonlyUint8Array;
};

/** Held-delivery receipt PDA. */
export async function findReceiptPda(
  seeds: ReceiptSeeds,
  config: { programAddress?: Address | undefined } = {},
): Promise<ProgramDerivedAddress> {
  const {
    programAddress = 'GyrTVV4hbcuzJuSz86FNq7K2UVAoSJQtcgHTVTz1hPPq' as Address<'GyrTVV4hbcuzJuSz86FNq7K2UVAoSJQtcgHTVTz1hPPq'>,
  } = config;
  return await getProgramDerivedAddress({
    programAddress,
    seeds: [
      getBytesEncoder().encode(new Uint8Array([114, 101, 99, 101, 105, 112, 116])),
      getAddressEncoder().encode(seeds.receiver),
      getAddressEncoder().encode(seeds.mint),
      getAddressEncoder().encode(seeds.sourceOwner),
      fixEncoderSize(getBytesEncoder(), 32).encode(seeds.uniqueNonce),
    ],
  });
}